fastapi_image_search/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
transformers==4.44.2
torch
faiss-cpu==1.8.0.post1
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.9
